# openapi.yaml
openapi: 3.0.3
info:
  title: Liquor Distribution OS API
  version: 1.0.0
  description: |
    REST API for the Liquor Distribution Operating System.
    Covers customers, orders, invoices, collections, deliveries, inventory, and sync.
    All mutating endpoints require an `Idempotency-Key` header.
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging.api.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      parameters:
        - { in: query, name: customer_id, schema: { type: string } }
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
      responses: { '200': { description: ok } }
    post:
      tags: [Orders]
      summary: Create order (runs credit check inline)
      parameters:
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderInput' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderCreateResult' }
  /orders/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Orders]
      responses: { '200': { description: ok } }
...
  /orders/{id}/cancel:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Orders]
      responses: { '200': { description: ok } }

  # ───────────────────────────── INVOICES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderInput:
      type: object
      required: [customer_id, lines]
      properties:
        customer_id: { type: string }
        notes: { type: string }
        lines:
          type: array
          items:
            type: object
            required: [product_id, qty]
            properties:
              product_id: { type: string }
              qty: { type: integer }
              unit_price: { type: number, description: "optional; otherwise from price list" }
              discount_pct: { type: number }
    OrderCreateResult:
      type: object
      properties:
        order_id: { type: string }
        status: { type: string, enum: [confirmed, held, rejected] }
        credit: { $ref: '#/components/schemas/CreditDecision' }
    CreditDecision:
      type: object
      properties:
        decision: { type: string, enum: [approve, hold, reject] }
        available_credit: { type: number }
        order_total: { type: number }
        risk_score: { type: number }
        reasons: { type: array, items: { type: string } }
...